pyproject.toml
[project]
name = "interact2025"
version = "0.1.0"
description = "Eliciting User-Defined Mid-Air Hand Gestures for Hybrid Meeting Platform Control: Results, Insights, and Design Implications"
authors = [
    {name = "Elizabete Munzlinger, Fabricio Batista Narcizo, Renata Briet, Mario Tadashi Shimanuki, Ted Vucurevich, and Dan Witzner Hansen",email = "elizabete@elizabete.com.br"}
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "transformers (>=4.52.3,<5.0.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "openai (>=1.82.1,<2.0.0)",
    "google-genai (>=1.17.0,<2.0.0)",
    "ollama (>=0.5.0,<0.6.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "fastdtw (>=0.3.4,<0.4.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
